requirements.txt
# Core ML Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Image Processing
Pillow>=9.0.0

# GUI and Visualization
matplotlib>=3.5.0
tkinter  # Usually included with Python

# Web Requests
requests>=2.25.0

# Environment Variables
python-dotenv>=0.19.0

# Optional for advanced features
scipy>=1.7.0  # For median filter defense

# Development/Testing (optional)
pytest>=6.0.0
jupyter>=1.0.0
